+++ pkg/cli/cc/ports.go
@@ -1,17 +1,15 @@
 package cc
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"net"
 	"time"
 
 	"github.com/rs/zerolog/log"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 )
+
+var errTimeout = errors.New("timeout")
 
 type portStatus int
 
@@ -81,8 +79,6 @@ func PortStatus(listenAddress string) portStatus {
 	return PortClosed
 }
 
-var errTimeout = errors.New("timeout")
-
 func Retry(timeout, interval time.Duration, f func() error) (err error) {
 	err = errTimeout
 	for t := time.After(timeout); ; {
@@ -101,46 +97,4 @@ func Retry(timeout, interval time.Duration, f func() error) (err error) {
 			time.Sleep(interval)
 		}
 	}
-}
-
-// ServiceHealthStatus adopted from grpc-health-probe cli implementation
-// https://github.com/grpc-ecosystem/grpc-health-probe/blob/master/main.go.
-func ServiceHealthStatus(service string) bool {
-	addr := "localhost:9494"
-	connTimeout := time.Second * 30
-	rpcTimeout := time.Second * 30
-
-	bCtx := context.Background()
-	dialCtx, dialCancel := context.WithTimeout(bCtx, connTimeout)
-	defer dialCancel()
-
-	dialOpts := []grpc.DialOption{
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
-	}
-
-	conn, err := grpc.DialContext(dialCtx, addr, dialOpts...)
-	if err != nil {
-		return false
-	}
-	defer conn.Close()
-
-	rpcCtx, rpcCancel := context.WithTimeout(bCtx, rpcTimeout)
-	defer rpcCancel()
-
-	if err := Retry(rpcTimeout, time.Millisecond*100, func() error {
-		resp, err := healthpb.NewHealthClient(conn).Check(rpcCtx, &healthpb.HealthCheckRequest{Service: service})
-		if err != nil {
-			return err
-		}
-
-		if resp.GetStatus() != healthpb.HealthCheckResponse_SERVING {
-			return fmt.Errorf("gRPC endpoint not SERVING")
-		}
-
-		return nil
-	}); err != nil {
-		return false
-	}
-
-	return true
 }
